RDRP-780: Tests amended
